Move styling code
