+++ Cargo.toml
@@ -4,27 +4,20 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-geotiff = "0.0.2"
-warp = "0.3"
-tokio = { version = "1", features = ["full"] }
-image = "0.25.5"
-geo-types = "0.7.15"
-# gdal = { version = "0.17.1", features = ["bindgen"] }
-rust-s3 = "0.35.1"
-anyhow = "1.0.95"
-georaster = "0.2.0"
-tile-grid = "0.6.2"
-dotenvy = "0.15.7"
-serde = "1.0.217"
+anyhow = "1.0.97"
 axum = "0.8.1"
-redis = { version = "0.28.2", features = ["tokio-comp"] }
-proj4rs = "0.1.5"
-serde_json = "1.0.138"
-log = "0.4.25"
+dotenvy = "0.15.7"
 gdal = { version = "0.17.1", features = ["bindgen"] }
 gdal-sys = { version = "0.10.0", features = ["bindgen"] }
+geo-types = "0.7.15"
+georaster = "0.2.0"
+geotiff = "0.0.2"
+image = "0.25.5"
+log = "0.4.26"
+proj4rs = "0.1.5"
+redis = { version = "0.29.1", features = ["tokio-comp"] }
 retry = "2.0.0"
-tokio-retry = "0.3.0"
+rust-s3 = "0.35.1"
 sea-orm = { version = "1.1.7", features = [
     "sqlx-postgres",
     "runtime-tokio-rustls",
@@ -33,3 +26,9 @@ sea-orm = { version = "1.1.7", features = [
     "with-chrono",
     "with-json",
 ], default-features = false }
+serde = "1.0.219"
+serde_json = "1.0.140"
+tile-grid = "0.6.2"
+tokio = { version = "1.44.1", features = ["full"] }
+tokio-retry = "0.3.0"
+warp = "0.3"
